% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brackets.R
\name{brackets}
\alias{brackets}
\title{Brackets}
\usage{
brackets(authority.in)
}
\arguments{
\item{authority.in}{authority as returned by UKSI}
}
\value{
text with brackets
}
\description{
Returns a consistent layout for authority without changing base data.
Deprecated:
Now does not change anything as the
position of the brackets encode species status.
}
\examples{

"Davies, 1979"


}
