% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.dates.R
\name{Clean.dates}
\alias{Clean.dates}
\title{Clean.Dates}
\usage{
Clean.dates(x)
}
\arguments{
\item{x}{A vector of dates}
}
\value{
A multicolumn vector of the same length as x with the following values:

"YYMMDD", "week", "YYYY", "DD", "MM", "MMDD", "DOY"
}
\description{
Converts dates in inconsistent formats into a consistent format.
It now uses string functions.
}
\examples{
my.cleaned.dates <- Clean.dates(c("1916-01-01", "2000 to 2010",
"July 1996 to August 1997"  ))

}
