% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.pdf.files.R
\name{compress.pdf.files}
\alias{compress.pdf.files}
\title{Compress PDF Files}
\usage{
compress.pdf.files(project.path, is.recursive = FALSE)
}
\arguments{
\item{project.path}{Usually the project directory.}

\item{is.recursive}{Recursively search for PDFs if TRUE.}
}
\value{
Nothing
}
\description{
Compress PDF files built with markdown.
}
