% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.contributors.R
\name{eco.contributors}
\alias{eco.contributors}
\title{Contributors To Report}
\usage{
eco.contributors(my.data.ss)
}
\arguments{
\item{my.data.ss}{Search data for names of contributors}
}
\value{
List of contributors and clipboard text.
}
\description{
Contributors To Report
}
