% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBNG.to.Lat.Long.R
\name{GBNG.to.Lat.Long}
\alias{GBNG.to.Lat.Long}
\title{GBNG.to.Lat.Long}
\usage{
GBNG.to.Lat.Long(GBNG)
}
\arguments{
\item{GBNG}{Must be a list of grid references. Spaces are ignored}
}
\value{
Latitiude and Longitude
}
\description{
This function takes a GB national Grid reference and returns Lattitude and
Longitude
}
\examples{
GBNG.to.Lat.Long(list("SK8903708058","SK8882008116", "SK 5183 1879"))

}
