% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_excel_data.R
\name{get_excel_data}
\alias{get_excel_data}
\title{Get Excel Data}
\usage{
get_excel_data(data_raw_directory = "data-source", columns_list)
}
\arguments{
\item{data_raw_directory}{Source of the raw data as a list of files.}

\item{columns_list}{A list of column mappings.}
}
\value{
A count of the number of files read. Output saved into data-clean
directory.
}
\description{
This function discovers Survey data in a directory (data-raw by default) and
processess it by extracting a consistent series of columns and placing them in
a data-clean directory. This function may be used in conjucntion with other
related functions to extract data from other sources.
}
