% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_to_bb.R
\name{ggmap_to_bbox}
\alias{ggmap_to_bbox}
\title{GGMAP Bounding Box to Bounding Box}
\usage{
ggmap_to_bbox(ggmap_object, xcoord, ycoord, zoom_factor = 1)
}
\arguments{
\item{ggmap_object}{ggmap object from which the bounding box is derived.}

\item{xcoord}{x cordinates in the non specific units.}

\item{ycoord}{y coordinates in non specific units.}

\item{zoom_factor}{Zoom factor. Defaults to 1. Bigger numbers zoom in.}
}
\value{
A bounding box
}
\description{
The ggmap bounding box needs transforming so it may be used to guid visual
exploration of data. This function helps by recreating bounding boxes derived
from existing maps. The user can enter bew coordinates to centre the bounding
box, then increase the zoom value to centre on the data.
}
