% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpegs.to.report.format.R
\name{jpegs.to.report.format}
\alias{jpegs.to.report.format}
\title{Jpegs To Report Format}
\usage{
jpegs.to.report.format(jpg.path = "_records")
}
\arguments{
\item{jpg.path}{Path to directory with jpg files relative to project
directory}
}
\value{
nothing to return
}
\description{
Resize and rename jpeg files to jpg with 1000px height.
}
