% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.pdf.R
\name{preview.pdf}
\alias{preview.pdf}
\title{Preview.pdf}
\usage{
preview.pdf()
}
\value{
Nothing returned
}
\description{
Generates a preview of the currently active bookdown chapter and compresses
the PDF file.
}
