% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_UNNG_Letter.R
\name{sf_to_UKNGR_Letter}
\alias{sf_to_UKNGR_Letter}
\title{Simple Feature Coordinates to UKNG Letter}
\usage{
sf_to_UKNGR_Letter(sf_coords)
}
\arguments{
\item{sf_coords}{Simple features object.}
}
\value{
A dataframe of tetrads the same length as the input.
}
\description{
Accepts a Simple Features object and returns the enclosing tetrad with
leading letters.
}
