% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKSI.look.up.taxon.R
\name{UKSI.look.up.taxon}
\alias{UKSI.look.up.taxon}
\title{Look Up Taxon From UKSI}
\usage{
UKSI.look.up.taxon(taxon)
}
\arguments{
\item{taxon}{character: species to lookup One at a time}
}
\value{
tibble
}
\description{
Look Up Taxon From UKSI
}
