% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_long_to_GBNG.R
\name{lat_long_to_GBNG}
\alias{lat_long_to_GBNG}
\title{Lattitude Longitude To UKNG}
\usage{
lat_long_to_GBNG(lattitude, longitude)
}
\arguments{
\item{lattitude}{as it says}

\item{longitude}{as it says}
}
\value{
GBNG value
}
\description{
Lattitude Longitude To UKNG
}
\examples{
 lat_long_to_GBNG(52.665137, -0.6906808)

}
