% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.record.R
\name{new.record}
\alias{new.record}
\title{new.record}
\usage{
new.record(
  index.reference,
  record.date = "Date",
  place.name = "Place Name",
  grid.ref = "Grid Reference",
  recorder.name = "Recorder Name",
  record.notes = "Record Notes",
  determiner = "Determiner Name",
  order = "Lepidoptera"
)
}
\arguments{
\item{index.reference}{A unique record identifier.}

\item{record.date}{ISO Date.}

\item{place.name}{Place name.}

\item{grid.ref}{Grid ref.}

\item{recorder.name}{Recorder Name.}

\item{record.notes}{Record notes.}

\item{determiner}{The determiner.}

\item{order}{Taxon order.}
}
\value{
Nothing returned
}
\description{
new.record
}
