% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.record.R
\name{new.record}
\alias{new.record}
\title{new.record}
\usage{
new.record(index.reference, datum.attributes, data.model.path = NULL)
}
\arguments{
\item{index.reference}{A unique record identifier.}

\item{datum.attributes}{Named list of attributes to use in template}

\item{data.model.path}{Path to external template}
}
\value{
Nothing returned
}
\description{
new.record
}
