% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.record.R
\name{new.record}
\alias{new.record}
\title{new.record}
\usage{
new.record(
  index.reference,
  record.date = "Date",
  place.name = "Place Name",
  grid.ref = "Grid Reference",
  recorder.name = "Recorder Name",
  record.notes = "Record Notes",
  determiner = "Determiner Name",
  order = "Lepidoptera",
  family = "Family",
  sub.family = "Sub-Family",
  genus = "Genus",
  taxon = "Taxon",
  log.number = "Classification Log Number",
  Bradley = "Bradley Fletcher Number",
  common.name = "Common Name",
  gender = c("Male, Female")
)
}
\arguments{
\item{index.reference}{A unique record identifier.}

\item{record.date}{ISO Date.}

\item{place.name}{Place name.}

\item{grid.ref}{Grid ref.}

\item{recorder.name}{Recorder Name.}

\item{record.notes}{Record notes.}

\item{determiner}{The determiner.}

\item{order}{Taxon order.}

\item{family}{Family.}

\item{sub.family}{Sub.family often used for Lepidoptera.}

\item{genus}{Genus without species name.}

\item{taxon}{Genus and species name.}

\item{log.number}{Log number for putting species in taxonomic order.}

\item{Bradley}{Bradley Fletcher Log number.}

\item{common.name}{Vernacular name.}

\item{gender}{Male or Female?}
}
\value{
Nothing returned
}
\description{
new.record
}
